Major Changes in Generation of AST nodes from the configuration file
The ast nodes file is now a .cfg. (Edited setup.py for that)
No longer have typed nodes (Python's Dynamic!!, and I never wanted to implement them, wonder how I got the idea).
The sources generated are cleaner.
Better repr.
Now raises errors that are defined in the module.
Better test coverage. (Not as comprehensive as we'd want)