[REFACTOR, TESTS] Change mechanism for data-driven-tests
Change to a much nicer decorater based data-driven-testing mechanism.

Now, instead of 2 functions (one for test-logic and one
containing/yielding tests), we will use a single decorated function and
an iterable. The iterable is passed to the decorator along with other
description-related arguments. The decorator does some magic (similar to
the one we did earlier) and returns a generator function that yields out
tests.

This approach is cleaner IMO as it makes the "data" of the test separate
from the logic while using only one function.