[TESTS][REWORK] Rewrite the data-driven-tests decorator
This is a major rework of the data driven tests infrastructure.

Earlier, all the tests were in only python code. Often I needed a
template function that I would call from a test-function. This doesn't
really scale nicely.

Now, the decorator adds support for loading from external YAML files.
Finally the data is moved out of the test file and has a nice place to
be in. Also, the data and behaviour better separated, letting me focus
on one at a time (if that's possible).

Plus YAML is a nicer format for data, it's definitely better than Python
for it. I'm using Python objects in tests as "data", I'm probably going
to evaluate string in a sandboxed context...

If I end up doing any more work on this, I'll first make a PyPI package
for this. Because only then will it justify the unethical amount of time
I've spent on this.